Refactor function labels to python standard
